Add/warehouse (#15)
* [REFACTOR] controllers now return HttpStatus instead of sting

* [ADD] Warehouse entity and [REFACTOR] PurchaseService and AlcoholService

* [FIX] PurchaseServiceTest

* [FIX] error codes and implements test for check correctness

* [FIX] test which use warehouse

* [ADD] test WarehouseTest class

* [REFACTOR] Warehouse class with annotation getter and setter instead of data

* [REFACTOR] returned response code

* [REFACTOR] exceptions to AppBaseException

* [FIX] sonar problems

+++ src/main/java/com/zzpj/backend/api/v1/AlcoholController.java
@@ -2,9 +2,12 @@ package com.zzpj.backend.api.v1;
 
 import com.zzpj.backend.dto.AlcoholDTO;
 import com.zzpj.backend.entities.Alcohol;
+import com.zzpj.backend.exceptions.AppBaseException;
 import com.zzpj.backend.mappers.AlcoholMapper;
 import com.zzpj.backend.services.interfaceses.AlcoholServiceLocal;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -23,38 +26,50 @@ public class AlcoholController {
     }
 
     @GetMapping
-    public List<Alcohol> getAll() {
-        return alcoholService.getAllAlcohols();
+    public ResponseEntity<List<Alcohol>> getAll() {
+        return new ResponseEntity<>(alcoholService.getAllAlcohols(), HttpStatus.OK);
     }
 
     @GetMapping("{id}")
-    public Alcohol get(@PathVariable Long id) {
+    public ResponseEntity<Alcohol> get(@PathVariable Long id) {
         Optional<Alcohol> alcohol = alcoholService.getAlcohol(id);
-        return alcohol.orElseGet(Alcohol::new);
+        return new ResponseEntity<>(alcohol.orElseGet(Alcohol::new), HttpStatus.OK);
     }
 
     @PostMapping
-    public String add(@RequestBody AlcoholDTO alcoholDTO) {
+    public ResponseEntity<String> add(@RequestBody AlcoholDTO alcoholDTO) {
         AlcoholMapper alcoholMapper = new AlcoholMapper();
         Alcohol alcohol = alcoholMapper.mapAlcoholDTOToAlcohol(alcoholDTO);
-        alcohol.setId(null);
-
-        alcoholService.addAlcohol(alcohol);
-        return "Success";
+        try {
+            if(alcohol.getName() == null) throw new AppBaseException();
+            alcoholService.addAlcohol(alcohol);
+        } catch (AppBaseException e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @PutMapping
-    public String edit(@RequestBody AlcoholDTO alcoholDTO) {
+    public ResponseEntity<String> edit(@RequestBody AlcoholDTO alcoholDTO) {
         AlcoholMapper alcoholMapper = new AlcoholMapper();
         Alcohol alcohol = alcoholMapper.mapAlcoholDTOToAlcohol(alcoholDTO);
-
-        alcoholService.editAlcohol(alcohol);
-        return "Success";
+        try {
+            if(alcohol.getId() != null && alcohol.getId() < 0) throw new AppBaseException();
+            alcoholService.editAlcohol(alcohol);
+        } catch (AppBaseException e) {
+            return new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
+        }
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping("{id}")
-    public String delete(@PathVariable Long id) {
-        alcoholService.deleteAlcohol(id);
-        return "Success";
+    public ResponseEntity<String> delete(@PathVariable Long id) {
+        try {
+            if(id < 0) throw new AppBaseException();
+            alcoholService.deleteAlcohol(id);
+        } catch (AppBaseException e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
